Refactoring and reorganization
